+++ src/SFA.DAS.DataAnalytics.Tests/Project/ImplicitUsing.cs
@@ -1,0 +1,7 @@
+﻿global using SFA.DAS.ConfigurationBuilder;
+global using SFA.DAS.FrameworkHelpers;
+global using System;
+global using System.Collections.Generic;
+global using NUnit.Framework;
+global using SFA.DAS.DataAnalytics.Tests.Project.Helpers;
+global using TechTalk.SpecFlow;
