+++ src/SFA.DAS.DataAnalytics.Tests/appsettings.Environment.json
@@ -1,4 +1,4 @@
 {
   "local_EnvironmentName": "PP",
-  "ProjectName": "TestDataCreation"
+  "ProjectName": "DataAnalytics"
 }
